Change structure, reformat code, move some logic to Game class
