+++ .gitignore
@@ -2,3 +2,5 @@
 /venvv
 /__pycache__/
 /.idea
+/levels/__pycache__
+/models/__pycache__
