+++ levels/Level.py
@@ -1,0 +1,17 @@
+class Level:
+    def __init__(self, player):
+        self.set_of_squares = set()
+        self.set_of_obstacles = set()
+        self.player = player
+
+    def update(self):
+        for p in self.set_of_squares:
+            p.update()
+        for p in self.set_of_obstacles:
+            p.update()
+
+    def draw(self, surface):
+        for p in self.set_of_squares:
+            p.draw(surface)
+        for p in self.set_of_obstacles:
+            p.draw(surface)
