+++ models/Monster.py
@@ -1,2 +1,4 @@
+# Klasa przeciwnika
 class Monster:
     def __init__(self):
+        pass
